refactor: UrlParser -> GithubParser로 이름 변경
- 추후 경로 이외의 파싱이 들어가 변경합니다
